Second half of pulling updates from branch
Pulling from tag works as well (at least for git; there's a lot of work to still be done for svn) though commit hash is untested. Speaking of work to be done, should probably also pull a big code clean-up sonn. Regardless, this is probably the  point where I can start the task of adding engines to the DB again.